refactor test helpers into single file, adjust coverage config to ignore helpers
