+++ src/helpers.ts
@@ -55,6 +55,8 @@ export function getTypeAndConversation(body: any): { type?: IncomingEventType, c
   return {};
 }
 
+/* istanbul ignore next */
+
 /** Helper that should never be called, but is useful for exhaustiveness checking in conditional branches */
 export function assertNever(x: never): never {
   throw new Error(`Unexpected object: ${x}`);
